+++ Changelog.md
@@ -5,6 +5,7 @@ OpenCore Changelog
 - Fixed CPU frequency calculation on AMD 10h family
 - Swapped the position of Shutdown and Restart buttons to better match recent macOS
 - Added `OC_ATTR_USE_REVERSED_UI` to allow access to previous default Shutdown and Restart button arrangement
+- Fixed intro animation getting stuck in OpenCanopy if an entry which returns to menu is selected before animation ends
 
 #### v0.9.7
 - Updated recovery_urls.txt